+++ lib/ws_servers/api/handlers/on_change_mode.js
@@ -1,5 +1,5 @@
 const sendAuthenticated = require('../send_authenticated')
-const sendDataAgain = require('../send_data_again')
+const resendSnapshots = require('../snapshots/send_all')
 const sendError = require('../../../util/ws/send_error')
 
 module.exports = async (server, ws, msg) => {
@@ -15,5 +15,5 @@ module.exports = async (server, ws, msg) => {
   await sendAuthenticated(server, ws, { mode, dmsScope })
 
   // TODO: send only if conn is not new
-  await sendDataAgain(ws)
+  await resendSnapshots(ws)
 }
